requirements.txt
loguru
chardet
pyside6
pysmb
paramiko
requests
py7zr
pycryptodome

pytest
coverage